Replace templates controllers to external and internal.
