+++ DevelopGitHub/GitHubExtension/GitHubExtension.Templates/GitHubExtension.Templates.csproj
@@ -53,8 +53,8 @@
       <HintPath>..\packages\Microsoft.AspNet.Identity.EntityFramework.2.2.1\lib\net45\Microsoft.AspNet.Identity.EntityFramework.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\packages\Newtonsoft.Json.6.0.4\lib\net45\Newtonsoft.Json.dll</HintPath>
+    <Reference Include="Newtonsoft.Json, Version=8.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.8.0.3\lib\net45\Newtonsoft.Json.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="SimpleInjector, Version=3.1.2.0, Culture=neutral, PublicKeyToken=984cb50dea722e99, processorArchitecture=MSIL">
@@ -92,8 +92,8 @@
     <Compile Include="CommunicationModels\TemplatesModel.cs" />
     <Compile Include="CommunicationModels\CreateUpdateTemplateModel.cs" />
     <Compile Include="Constants\RouteConstants.cs" />
-    <Compile Include="Controllers\TemplatesCommandController.cs" />
-    <Compile Include="Controllers\TemplatesQueryController.cs" />
+    <Compile Include="Controllers\TemplatesExternalController.cs" />
+    <Compile Include="Controllers\TemplatesInternalController.cs" />
     <Compile Include="ExtensionMethods\HttpClientExtension.cs" />
     <Compile Include="ExtensionMethods\HttpRequestMessageExtension.cs" />
     <Compile Include="ExtensionMethods\HttpResponseMessageExtension.cs" />
